// Lab3/pzpi-22-3-protsenko-artem-lab3/src/pages/router/routes.enums.ts
export enum ROUTES {
    HOME = "/",
    REGISTRATION_PAGE = "/register",
    LOGIN_PAGE = "/login",
    USER_PROFILE = "/profile/:id",
    VERIFICATION_PAGE = "/verification/:email",
    CREATE_PET_PAGE = "/create-pet",
    PET_LIST_PAGE = "/pet-list/:id",
    PET_PAGE = "/pet/:id",
    UPDATE_PET_PAGE = "/update-pet/:id",
    ADD_MEAL = "/add-meal/:id",
    EDIT_MEAL = "/update-meal/:id",
    DB_ADMIN_PAGE = "/db-admin-panel",
    SYSTEM_ADMIN_PAGE = "/system-admin-panel",
    LOCAL_ADMIN_PAGE = "/local-admin-panel",
    HEALTH_ANALYSIS_LIST_PAGE = "/health-analysises/:id",
    HEALTH_ANALYSIS_PAGE = "/health-analysis/:id",
}